requirements.txt
praw
pandas
numpy
configparser
s3fs
apache-airflow-providers-amazon
